Split resolver local absence tests
